[proliferate] fix pending shell repo shape
